use double for rendering

+++ src/main/kotlin/dev/nyon/moredetails/util/PoseStackExtensions.kt
@@ -3,8 +3,8 @@ package dev.nyon.moredetails.util
 import com.mojang.blaze3d.vertex.PoseStack
 import net.minecraft.client.gui.GuiComponent
 
-fun PoseStack.verticalLine(x: Int, y: Int, height: Int, color: Int = 0x000000) =
-    GuiComponent.fill(this, x, y, x + 1, y + height, color)
+fun PoseStack.verticalLine(x: Double, y: Double, height: Double, color: Int = 0x000000) =
+    GuiComponent.fill(this, x.toInt(), y.toInt(), (x + 1).toInt(), (y + height).toInt(), color)
 
-fun PoseStack.horizontalLine(x: Int, y: Int, width: Int, color: Int = 0x000000) =
-    GuiComponent.fill(this, x, y, x + width, y + 1, color)
+fun PoseStack.horizontalLine(x: Double, y: Double, width: Double, color: Int = 0x000000) =
+    GuiComponent.fill(this, x.toInt(), y.toInt(), (x + width).toInt(), (y + 1).toInt(), color)
